Update User model to use soft deletes and modify users table
The User model has been updated to use the SoftDeletes trait, adding a "deleted\_at" column to the users table. The "name" column in the users table has been changed to "full\_name" and a new "username" column has been added, which is unique.

The API routes have been updated to include a version prefix (v1) and to include a test route and authentication routes for registration, login, and logout.